settings code moved to settings.py
