+++ Code/weatherApp/templates/index.html.jinja
@@ -19,7 +19,7 @@
     <button onclick="window.location.href = '{{ url_for('views.map') }}';" class="button-style">Map</button>
     <button onclick="window.location.href = '{{ url_for('views.weather_summary') }}';" class="button-style">Weather Summary</button>
     <button onclick="window.location.href = '{{ url_for('views.graph') }}';" class="button-style">Graphs</button>
-    <button onclick="window.location.href = '{{ url_for('views.settings') }}';" class="button-style">Settings</button>
+    <button onclick="window.location.href = '{{ url_for('settings.settings') }}';" class="button-style">Settings</button>
     <button onclick="window.location.href = '{{ url_for('auth.admin_dashboard') }}';" class="button-style">Admin Dashboard</button>
   </div>
