+++ Code/weatherApp/templates/settings.html.jinja
@@ -14,7 +14,7 @@
 {% block content %}
   <h2>Update Profile</h2>
 
-  <form id="userSettingsForm" method="post" action="{{ url_for('views.settings') }}">
+  <form id="userSettingsForm" method="post" action="{{ url_for('settings.settings') }}">
     <div style="display: flex; align-items: center;">
       <label for="emailList">Email List:</label>
       <input type="checkbox" id="emailList" name="emailList" {% if user.emailList %} checked {% endif %}>
